Improve `UniqueTest` and move `UniqueWhereNotNullMacro` to `Rhetos.DatabaseGenerator.DefaultConcepts` to solve referencing
